Revert "refactor(core): move workspace sql under worktree"
This reverts commit 3c2005a244f0e3bbc20d58b40b2d3b8073763e76.
