+++ packages/core/test/workspace.test.ts
@@ -4,7 +4,7 @@ import { Database } from "@opencode-ai/core/database/database"
 import { makeMemoryDriver } from "@opencode-ai/core/environment/index"
 import { Workspace } from "@opencode-ai/core/workspace"
 import { WorkspaceDriver } from "@opencode-ai/core/workspace/driver"
-import { WorkspaceTable } from "@opencode-ai/core/worktree/sql"
+import { WorkspaceTable } from "@opencode-ai/core/workspace/sql"
 import { LayerNode } from "@opencode-ai/util/effect/layer-node"
 import { eq } from "drizzle-orm"
 import { Effect } from "effect"
